requirements.txt
torch>=1.9.0
numpy>=1.21.0
gymnasium>=0.28.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0
tensorboard>=2.8.0
tqdm>=4.64.0
pandas>=1.4.0
scipy>=1.8.0
wandb>=0.15.0 